requirements.txt
# Crypto Monitor v8.3 - Unified Dependencies
# 统一依赖文件

# ==== Core ====
redis==5.0.3
pyyaml==6.0.1
python-dotenv==1.0.0

# ==== Async / HTTP ====
aiohttp==3.9.1
httpx==0.25.2
websockets==12.0

# ==== JSON ====
orjson==3.9.10

# ==== Node A: Exchange Monitor ====
# (aiohttp, websockets already included)

# ==== Node B: Blockchain + Twitter + News ====
web3==6.11.3
tweepy==4.14.0
feedparser==6.0.10

# ==== Node C: Korea + Telegram ====
python-telegram-bot==20.7
telethon==1.34.0

# ==== Fusion / Dashboard ====
flask==3.0.0
flask-cors==4.0.0

# ==== Optional: Development ====
# pytest==7.4.3
# black==23.12.1

